+++ test/stepper_motor/journey/step_definition_test.rb
@@ -17,7 +17,7 @@ class StepDefinitionTest < ActiveSupport::TestCase
   end
 
   test "passes any additional options to the step definition" do
-    step_def = StepperMotor::Step.new(name: "a_step", seq: 1, on_exception: :reattempt!)
+    step_def = StepperMotor::Step.new(name: "a_step", on_exception: :reattempt!)
     assert_extra_arguments = ->(**options) {
       assert options.key?(:extra)
       # Return the original definition